RayCast -> SphereCast
Testing showed that the raycast caused an issue. This issue was the ray is a 1D object that originates from the centre of the player. This means if the player stood on the edge of a object, the ray could miss and report back that the player was not grounded as the ray was hitting an object greater than the tolerance away. By making this RayCast a SphereCast, I can project a ray that is the width of the player instead. 
The tolerance was also changed to 0.11, as the player was floating by accident with 1.11.

Gitignore updated to ignore editor layout.
Values in script now align with editor set ones.