SphereCast Size Adjust
The sphereCast is now visible in the editor, this showed me that the SphereCast was twice the size of the player. This was because I assumed the SphereCast was being fed a DIAMETER of 1. However it was the radius. This led to the spherecast not matching the size of the player.